[mq2-driver] - make read take u16 instead of u8

+++ lib/mq2/src/mq2.c
@@ -105,7 +105,7 @@ error_type_t mq2_deinit(mq2_t* mq2_object){
     return OK;
 }
 
-error_type_t mq2_analog_read(mq2_t *mq2_object, uint8_t *value_ptr){
+error_type_t mq2_analog_read(mq2_t *mq2_object, uint16_t *value_ptr){
     if(mq2_object == NULL || value_ptr == NULL)return NULL_PARAMETER;
     if(mq2_object->initialized != true) return INVALID_STATE;
     if(!(mq2_object->mode == MQ2_ANALOG_ONLY || mq2_object->mode == MQ2_HYBRID )){
@@ -114,7 +114,7 @@ error_type_t mq2_analog_read(mq2_t *mq2_object, uint8_t *value_ptr){
     *value_ptr = analogRead(mq2_object->analog_pin_number);
     return OK;
 }
-error_type_t mq2_digital_read(mq2_t *mq2_object, uint8_t *value_ptr){
+error_type_t mq2_digital_read(mq2_t *mq2_object, uint16_t *value_ptr){
     if(mq2_object == NULL || value_ptr == NULL)return NULL_PARAMETER;
     if(mq2_object->initialized != true) return INVALID_STATE;
     if(!(mq2_object->mode == MQ2_DIGITAL_ONLY || mq2_object->mode == MQ2_HYBRID )){
